use `tune::.estimate_metrics()` to estimate metrics
Still tricky because the raw data usually does not encode factors as characters, so the transformation from numbers to factor levels depends on what approach the LLM takes to convert data to factor. That said, if the submission has the correct format, the column names of the submission is enough information to infer how the LLM chose to level the factor.

Using `.estimate_metrics()` allows us to lean on tune to decide which columns should be passed to the metric function.