revamp submission process
The assistant knows nothing about a "submission" process. Instead, the mock analyst has an `end_conversation()` tool that it can use to end a conversation once it receives affirmation from the assistant that it's done.

Then, once the conversation ends, the best workflow is extracted from predictive's global `the$experiments` deterministically and evaluated against unseen test data. This also means that the assistant now never needs to see the test data.

This _does_ mean that all transformations to data have to be encoded in the submitted recipe. This instruction will probably have to make it into predictive itself.

Related to #14 but does not close.